requirements.txt
anyio==3.5.0
asgiref==3.5.1
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.76.0
h11==0.13.0
idna==3.3
isort==5.10.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pydantic==1.9.0
requests==2.27.1
sniffio==1.2.0
starlette==0.18.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
python-dotenv==0.19.0
